+++ packages/framer-motion/src/render/VisualElement.ts
@@ -18,7 +18,6 @@ import { isMotionValue } from "../value/utils/is-motion-value"
 import { transformProps } from "./html/utils/transform"
 import {
     ResolvedValues,
-    VariantStateContext,
     VisualElementEventCallbacks,
     VisualElementOptions,
 } from "./types"
@@ -27,14 +26,12 @@ import {
     isControllingVariants as checkIsControllingVariants,
     isVariantNode as checkIsVariantNode,
 } from "./utils/is-controlling-variants"
-import { isVariantLabel } from "./utils/is-variant-label"
 import { updateMotionValuesFromProps } from "./utils/motion-values"
 import { resolveVariantFromProps } from "./utils/resolve-variants"
 import { warnOnce } from "../utils/warn-once"
 import { featureDefinitions } from "../motion/features/definitions"
 import { Feature } from "../motion/features/Feature"
 import type { PresenceContextProps } from "../context/PresenceContext"
-import { variantProps } from "./utils/variant-props"
 import { visualElementStore } from "./store"
 import { KeyframeResolver } from "./utils/KeyframesResolver"
 import { isNumericalString } from "../utils/is-numerical-string"
@@ -53,8 +50,6 @@ const propEventHandlers = [
     "LayoutAnimationStart",
     "LayoutAnimationComplete",
 ] as const
-
-const numVariantProps = variantProps.length
 
 /**
  * A VisualElement is an imperative abstraction around UI elements such as
@@ -653,34 +648,6 @@ export abstract class VisualElement<
             : this.parent
             ? this.parent.getClosestVariantNode()
             : undefined
-    }
-
-    getVariantContext(startAtParent = false): undefined | VariantStateContext {
-        if (startAtParent) {
-            return this.parent ? this.parent.getVariantContext() : undefined
-        }
-
-        if (!this.isControllingVariants) {
-            const context = this.parent
-                ? this.parent.getVariantContext() || {}
-                : {}
-            if (this.props.initial !== undefined) {
-                context.initial = this.props.initial as any
-            }
-            return context
-        }
-
-        const context = {}
-        for (let i = 0; i < numVariantProps; i++) {
-            const name = variantProps[i] as keyof typeof context
-            const prop = this.props[name]
-
-            if (isVariantLabel(prop) || prop === false) {
-                context[name] = prop
-            }
-        }
-
-        return context
     }
 
     /**
